// types/itemModel.ts
export interface ItemModel {
  ioi_id: string;
  ioi_it_id: string;
  ioi_name: string;
  ioi_desc: string | null;
  ioi_desc1: string | null;
  ioi_desc2: string | null;
  ioi_desc3: string | null;
  ioi_desc4: string | null;
  ioi_cur_code: string;
  ioi_unitprice: number;
  ioi_tax: number | null;
  ioi_disc: number | 0;
  ioi_photo1: string | null;
  ioi_photo2: string | null;
  ioi_photo3: string | null;
  ioi_photo4: string | null;
  ioi_photo5: string | null;
  ioi_photo6: string | null;
  ioi_photo7: string | null;
  ioi_photo8: string | null;
  ioi_photo9: string | null;
  ioi_video: string | null;
  ioi_remqty: number;
  ioi_inactive: boolean;
  ioi_masteritem: boolean;
  ioi_order: string;

  it_id: string;
  it_cmp_id: string;
  it_name: string;
  it_fa_name: string;
  it_group: string;
  it_bra_name: string;
  it_br_name: string | null;
  it_unit: string;
  it_cur_code: string;
  it_unitprice: number;
  it_vat: number;
  it_barcode: string | null;
  it_alertqty: number;
  it_di_name: string | null;
  it_inactive: boolean;
  it_size: string | null;
  it_color: string | null;
  it_code: string | null;
  it_points: number;
  it_div_name: string | null;
  it_image: string;
  it_profit: number;
  it_profrule: string | null;
  it_altname: string | null;
  it_wa_name: string | null;
  it_specialcode: boolean;
  it_userstamp: string;
  it_timestamp: string;
  it_commission: number;
  it_minprice: number;
  it_remqty: number;
  it_cost: number;
  it_lastsuppliername: string | null;
  it_lastsupplierprice: number;
  it_cashback: number;
  it_tax1: number;
  it_tax2: number;
  it_maxqty: number;
  it_order: string;
  it_pos: boolean;
  it_desc: string | null;
  it_type: string | null;
  it_online: boolean;
  it_digitalmenu: boolean;
  it_lineno: number;

  cur_code: string;
  cur_name: string;
  cur_order: number;
  cur_decimal: number;
  cur_round: number;
  cur_roundupdown: string;
  cur_newcode: string;
  cur_cmp_id: string;
  cur_maincurrency: boolean;

  br_name: string | null;
  br_parent: string | null;
  br_newname: string | null;
  br_cmp_id: string | null;

  fa_name: string;
  fa_parent: string;
  fa_group: string;
  fa_numberofyears: number | null;
  fa_codelen: number | null;
  fa_purchacc: string | null;
  fa_depacc: string | null;
  fa_depchargeacc: string | null;
  fa_salesacc: string | null;
  fa_costoffasoldacc: string | null;
  fa_periodicityofprovision: string | null;
  fa_newname: string;
  fa_cmp_id: string;

  rn: string;
}